feat: ロゴ画像のダークモード対応
ロゴ画像がカラー・モノクロで切り替わるように変更しました。
