+++ front/react-app/src/App.tsx
@@ -19,12 +19,27 @@ import {
   contactURL,
 } from "./constants/urls";
 import { Home } from "./Pages/Home";
-import { useContext } from "react";
+import { useContext, useEffect, useState } from "react";
 import { LoginContext } from "./providers/LoginFlagProvider";
 import { DarkMordToggle } from "./Organisms/DarkModeToggle";
 
 export const App = (): JSX.Element => {
   const { loginFlag } = useContext(LoginContext);
+  const [isDark, setIsDark] = useState<boolean>(false);
+  const htmlElement = document.getElementById('html');
+
+  useEffect(() => {
+    if (htmlElement?.classList.contains('dark')) {
+      setIsDark(true);
+    } else if (window.matchMedia('(prefers-color-scheme: dark)').matches === true) {
+      htmlElement?.classList.add('dark');
+    }
+  }, [htmlElement?.classList]) 
+
+  const toggleDarkClassForHtml = () => {
+    htmlElement?.classList.toggle('dark');
+    isDark ? setIsDark(false) : setIsDark(true);
+  }
   
   const LoginCheck = ({ component } : { component: JSX.Element }): JSX.Element => {
     if (loginFlag) {
@@ -38,7 +53,7 @@ export const App = (): JSX.Element => {
     <div className="fixed top-0 left-0 w-full h-full bg-[url('../../public/yoake.jpg')] bg-cover text-gray-700 dark:text-gray-200">
       <div className="w-2/3 mx-auto my-0">
         <Router>
-          <Header />
+          <Header isDark={ isDark } toggleDarkClassForHtml={ toggleDarkClassForHtml } />
           <div className="border w-2/3 h-2/3 border-gray-600 border-t-0 bg-sky-100 dark:bg-gray-600 fixed top-34 flex justify-center">
             <Routes>
               <Route path="/users" element={ <UsersIndex /> } />
@@ -53,7 +68,7 @@ export const App = (): JSX.Element => {
               <Route path="*" element={ <NotFound /> } />
             </Routes>
           </div>
-          <DarkMordToggle />
+          <DarkMordToggle isDark={ isDark } toggleDarkClassForHtml={ toggleDarkClassForHtml } />
         </Router>
       </div>
     </div>
